added a premede microchip project
As i found out, the proximity reporter is not what i needed and not a
viable solution for a ble beacon. The legaxy advertisement profile is
what i actually needed, because it doesn't connect and instead just
shouts into the void how i intended.
TL:DR: I don't like when it's assumed you know stuff but never explained